implement friend request system without actually saving friends rename search users page friend requests can be accepted and refused request system is handled on single page throw error on trying to cancel refused friend request handle ui hiding logic on frontend (FriendRequestState.vue)

+++ app/Http/Controllers/UserController.php
@@ -3,17 +3,17 @@
 namespace App\Http\Controllers;
 
 use App\Events\FriendRequestSent;
-use App\Http\Resources\FriendRequestResource;
 use App\Http\Resources\UserResource;
 use App\Models\FriendRequest;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\DB;
 use Illuminate\Support\Facades\Gate;
 use Inertia\Inertia;
 
 class UserController extends Controller
 {
-    public function search(Request $request)
+    public function usersAndFriendRequests(Request $request)
     {
         $input = $request->query('input', "");
 
@@ -23,6 +23,8 @@ class UserController extends Controller
 
             $users = User::with(['acceptableFriendRequests' => function ($query) use ($authUser) {
                 $query->where('sender_id', $authUser->getKey());
+            }])->with(['sentFriendRequests' => function ($query) use ($authUser) {
+                $query->where('recipient_id', $authUser->getKey());
             }])->when($input, function ($q) use ($input) {
                 $q->where('name', 'like', "%{$input}%");
             })->get();
@@ -30,21 +32,9 @@ class UserController extends Controller
             $users = [];
         }
 
-        return Inertia::render('SearchUsers', [
+        return Inertia::render('UsersAndFriendRequests', [
             'input' => $input,
             'users' => UserResource::collection($users),
-        ]);
-    }
-
-    public function sentFriendRequests()
-    {
-        /** @var User $user */
-        $user = auth()->user();
-
-        $friendRequests = $user->sentFriendRequests()->with('recipient')->get();
-
-        return Inertia::render('SentFriendRequests', [
-            'friendRequests' => FriendRequestResource::collection($friendRequests),
         ]);
     }
 
@@ -57,6 +47,12 @@ class UserController extends Controller
             'sender_id' => $user->getKey(),
             'recipient_id' => $recipient->getKey(),
         ])->exists();
+        if (!$exists) {
+            $exists = FriendRequest::where([
+                'sender_id' => $recipient->getKey(),
+                'recipient_id' => $user->getKey(),
+            ])->exists();
+        }
 
         if ($exists) {
             throw new \Exception ('Friend request already exists.');
@@ -76,9 +72,7 @@ class UserController extends Controller
 
         event(new FriendRequestSent($user, $recipient));
 
-        return redirect()->route('search-users', [
-            'input' => $request->query('input', ''),
-        ]);
+        return back();
     }
 
     public function cancelSentFriendRequest(Request $request, FriendRequest $friendRequest)
@@ -87,8 +81,45 @@ class UserController extends Controller
             abort(403);
         }
 
+        if ($friendRequest->refused) {
+            throw new \Exception ('Cannot cancel this request.');
+        }
+
         $friendRequest->delete();
 
-        return redirect()->route('sent-friend-requests');
+        return back();
+    }
+
+    public function acceptFriendRequest(Request $request, FriendRequest $friendRequest)
+    {
+        if (!Gate::allows('respond-to-friend-request', $friendRequest)) {
+            abort(403);
+        }
+
+        DB::beginTransaction();
+        try {
+            // Save friendRequest as a friend model/pivot
+
+            $friendRequest->delete();
+
+            DB::commit();
+        } catch (\Exception $e) {
+            DB::rollBack();
+            throw $e;
+        }
+
+        return back();
+    }
+
+    public function refuseFriendRequest(Request $request, FriendRequest $friendRequest)
+    {
+        if (!Gate::allows('respond-to-friend-request', $friendRequest)) {
+            abort(403);
+        }
+
+        $friendRequest->refused = true;
+        $friendRequest->save();
+
+        return back();
     }
 }
